CCM-14671: Make message plan campaign ID uneditable
