+++ frontend/src/components/molecules/PreviewTemplateDetails/PreviewTemplateDetails.module.scss
@@ -40,4 +40,3 @@
   margin: 1rem !important;
   max-width: initial;
 }
-
