+++ frontend/src/middleware.ts
@@ -10,8 +10,8 @@ const protectedPaths = [
   /^\/create-email-template$/,
   /^\/create-nhs-app-template$/,
   /^\/create-text-message-template$/,
-  /^\/delete-template\/[^/]+$/,
   /^\/delete-template-error\/[^/]+$/,
+  /^\/delete-template\/[^/]+$/,
   /^\/edit-email-template\/[^/]+$/,
   /^\/edit-letter-template\/[^/]+$/,
   /^\/edit-nhs-app-template\/[^/]+$/,
@@ -24,27 +24,27 @@ const protectedPaths = [
   /^\/letter-template-approved\/[^/]+$/,
   /^\/letter-template-submitted\/[^/]+$/,
   /^\/message-plans\/campaign-id-required$/,
-  /^\/message-plans\/choose-email-template\/[^/]+$/,
-  /^\/message-plans\/choose-email-template\/[^/]+\/preview-template\/[^/]+$/,
-  /^\/message-plans\/choose-british-sign-language-letter-template\/[^/]+$/,
   /^\/message-plans\/choose-british-sign-language-letter-template\/[^/]+\/preview-template\/[^/]+$/,
-  /^\/message-plans\/choose-large-print-letter-template\/[^/]+$/,
+  /^\/message-plans\/choose-british-sign-language-letter-template\/[^/]+$/,
+  /^\/message-plans\/choose-email-template\/[^/]+\/preview-template\/[^/]+$/,
+  /^\/message-plans\/choose-email-template\/[^/]+$/,
   /^\/message-plans\/choose-large-print-letter-template\/[^/]+\/preview-template\/[^/]+$/,
+  /^\/message-plans\/choose-large-print-letter-template\/[^/]+$/,
   /^\/message-plans\/choose-message-order$/,
-  /^\/message-plans\/choose-nhs-app-template\/[^/]+$/,
   /^\/message-plans\/choose-nhs-app-template\/[^/]+\/preview-template\/[^/]+$/,
-  /^\/message-plans\/choose-other-language-letter-template\/[^/]+$/,
+  /^\/message-plans\/choose-nhs-app-template\/[^/]+$/,
   /^\/message-plans\/choose-other-language-letter-template\/[^/]+\/preview-template\/[^/]+$/,
-  /^\/message-plans\/choose-standard-english-letter-template\/[^/]+$/,
+  /^\/message-plans\/choose-other-language-letter-template\/[^/]+$/,
   /^\/message-plans\/choose-standard-english-letter-template\/[^/]+\/preview-template\/[^/]+$/,
-  /^\/message-plans\/choose-text-message-template\/[^/]+$/,
+  /^\/message-plans\/choose-standard-english-letter-template\/[^/]+$/,
   /^\/message-plans\/choose-text-message-template\/[^/]+\/preview-template\/[^/]+$/,
+  /^\/message-plans\/choose-text-message-template\/[^/]+$/,
   /^\/message-plans\/create-message-plan$/,
   /^\/message-plans\/edit-message-plan\/[^/]+$/,
-  /^\/message-plans\/edit-message-plan-settings\/[^/]+$/,
   /^\/message-plans\/get-ready-to-move\/[^/]+$/,
   /^\/message-plans\/invalid$/,
   /^\/message-plans\/preview-message-plan\/[^/]+$/,
+  /^\/message-plans\/rename-message-plan\/[^/]+$/,
   /^\/message-plans\/review-and-move-to-production\/[^/]+$/,
   /^\/message-plans$/,
   /^\/message-templates$/,
@@ -59,18 +59,18 @@ const protectedPaths = [
   /^\/preview-submitted-text-message-template\/[^/]+$/,
   /^\/preview-text-message-template\/[^/]+$/,
   /^\/request-a-proof\/[^/]+$/,
-  /^\/review-and-approve-letter-template\/[^/]+$/,
   /^\/request-proof-of-template\/[^/]+$/,
+  /^\/review-and-approve-letter-template\/[^/]+$/,
   /^\/submit-email-template\/[^/]+$/,
   /^\/submit-letter-template\/[^/]+$/,
   /^\/submit-nhs-app-template\/[^/]+$/,
   /^\/submit-text-message-template\/[^/]+$/,
   /^\/text-message-template-submitted\/[^/]+$/,
   /^\/upload-british-sign-language-letter-template$/,
-  /^\/upload-other-language-letter-template$/,
   /^\/upload-large-print-letter-template$/,
   /^\/upload-letter-template\/client-id-and-campaign-id-required$/,
   /^\/upload-letter-template$/,
+  /^\/upload-other-language-letter-template$/,
   /^\/upload-standard-english-letter-template$/,
 ];
