+++ package.json
@@ -28,6 +28,9 @@
   },
   "name": "nhs-notify-web-template-management",
   "overrides": {
+    "@aws-sdk/xml-builder": {
+      "fast-xml-parser": "5.5.6"
+    },
     "fast-xml-parser": "^5.5.6",
     "flatted": "^3.4.0",
     "immutable": "^5.1.5",
@@ -38,10 +41,7 @@
       "react-is": "19.0.0"
     },
     "react": "^19.0.0",
-    "underscore": "^1.13.8",
-    "@aws-sdk/xml-builder": {
-      "fast-xml-parser": "5.5.6"
-    }
+    "underscore": "^1.13.8"
   },
   "scripts": {
     "build": "npm run create-env-file && npm run build --workspace frontend",
