+++ tests/test-team/pages/routing/index.ts
@@ -1,14 +1,14 @@
 export * from './campaign-id-required-page';
 export * from './choose-message-order-page';
-export * from './edit-message-plan-page';
 export * from './create-message-plan-page';
-export * from './edit-message-plan-settings-page';
+export * from './edit-message-plan-page';
+export * from './email';
 export * from './get-ready-to-move-page';
 export * from './invalid-message-plan-page';
-export * from './message-plans-page';
-export * from './preview-message-plan-page';
-export * from './review-and-move-to-production-page';
-export * from './email';
 export * from './letter';
+export * from './message-plans-page';
 export * from './nhs-app';
+export * from './preview-message-plan-page';
+export * from './rename-message-plan-page';
+export * from './review-and-move-to-production-page';
 export * from './sms';
