+++ tests/test-team/pages/routing/rename-message-plan-page.ts
@@ -1,9 +1,9 @@
 import { Locator, type Page } from '@playwright/test';
 import { TemplateMgmtBasePage } from 'pages/template-mgmt-base-page';
 
-export class RoutingEditMessagePlanSettingsPage extends TemplateMgmtBasePage {
+export class RoutingRenameMessagePlanPage extends TemplateMgmtBasePage {
   static readonly pathTemplate =
-    '/message-plans/edit-message-plan-settings/:messagePlanId';
+    '/message-plans/rename-message-plan/:messagePlanId';
 
   readonly submitButton: Locator;
 
@@ -11,22 +11,12 @@ export class RoutingEditMessagePlanSettingsPage extends TemplateMgmtBasePage {
 
   readonly nameFieldError: Locator;
 
-  readonly campaignIdSelector: Locator;
-
-  readonly campaignIdFieldError: Locator;
-
-  readonly singleCampaignIdElement: Locator;
-
   constructor(page: Page) {
     super(page);
     this.submitButton = page.getByTestId('submit-button');
     this.nameField = page.getByTestId('name-field');
-    this.campaignIdSelector = page.getByTestId('campaign-id-field');
-    this.singleCampaignIdElement = page.getByTestId('single-campaign-id');
 
     this.nameFieldError = page.locator('#name--error-message');
-
-    this.campaignIdFieldError = page.locator('#campaignId--error-message');
   }
 
   async clickSubmit() {
