+++ tests/test-team/template-mgmt-accessibility/routing.accessibility.spec.ts
@@ -15,10 +15,9 @@ import {
   RoutingChooseNhsAppTemplatePage,
   RoutingChooseOtherLanguageLetterTemplatePage,
   RoutingChooseStandardLetterTemplatePage,
-  RoutingEditMessagePlanPage,
   RoutingChooseTextMessageTemplatePage,
   RoutingCreateMessagePlanPage,
-  RoutingEditMessagePlanSettingsPage,
+  RoutingEditMessagePlanPage,
   RoutingGetReadyToMovePage,
   RoutingInvalidMessagePlanPage,
   RoutingMessagePlanCampaignIdRequiredPage,
@@ -31,6 +30,7 @@ import {
   RoutingPreviewOtherLanguageLetterTemplatePage,
   RoutingPreviewSmsTemplatePage,
   RoutingPreviewStandardLetterTemplatePage,
+  RoutingRenameMessagePlanPage,
   RoutingReviewAndMoveToProductionPage,
 } from 'pages/routing';
 import { getTestContext } from 'helpers/context/context';
@@ -326,17 +326,17 @@ test.describe('Routing', () => {
       )
     ));
 
-  test('Edit message plan settings', async ({ page, analyze }) =>
+  test('Rename message plan', async ({ page, analyze }) =>
     analyze(
-      new RoutingEditMessagePlanSettingsPage(page).setPathParam(
+      new RoutingRenameMessagePlanPage(page).setPathParam(
         'messagePlanId',
         draftRoutingConfigId
       )
     ));
 
-  test('Edit message plan settings - error', async ({ page, analyze }) =>
+  test('Rename message plan - error', async ({ page, analyze }) =>
     analyze(
-      new RoutingEditMessagePlanSettingsPage(page).setPathParam(
+      new RoutingRenameMessagePlanPage(page).setPathParam(
         'messagePlanId',
         draftRoutingConfigId
       ),
