+++ tests/test-team/template-mgmt-component-tests/template-protected-routes.component.spec.ts
@@ -10,7 +10,6 @@ import { RoutingChooseStandardLetterTemplatePage } from 'pages/routing/letter/ch
 import { RoutingChooseTextMessageTemplatePage } from 'pages/routing/sms/choose-sms-template-page';
 import { RoutingCreateMessagePlanPage } from '../pages/routing/create-message-plan-page';
 import { RoutingEditMessagePlanPage } from 'pages/routing/edit-message-plan-page';
-import { RoutingEditMessagePlanSettingsPage } from 'pages/routing/edit-message-plan-settings-page';
 import { RoutingGetReadyToMovePage } from 'pages/routing/get-ready-to-move-page';
 import { RoutingInvalidMessagePlanPage } from 'pages/routing/invalid-message-plan-page';
 import { RoutingMessagePlanCampaignIdRequiredPage } from '../pages/routing/campaign-id-required-page';
@@ -22,6 +21,7 @@ import { RoutingPreviewNhsAppTemplatePage } from 'pages/routing/nhs-app/preview-
 import { RoutingPreviewOtherLanguageLetterTemplatePage } from 'pages/routing/letter/preview-other-language-letter-template-page';
 import { RoutingPreviewSmsTemplatePage } from 'pages/routing/sms/preview-sms-template-page';
 import { RoutingPreviewStandardLetterTemplatePage } from 'pages/routing/letter/preview-standard-letter-page';
+import { RoutingRenameMessagePlanPage } from 'pages/routing/rename-message-plan-page';
 import {
   RoutingChooseBritishSignLanguageLetterTemplatePage,
   RoutingPreviewBritishSignLanguageLetterTemplatePage,
@@ -86,7 +86,6 @@ const protectedPages = [
   RoutingChooseTextMessageTemplatePage,
   RoutingCreateMessagePlanPage,
   RoutingEditMessagePlanPage,
-  RoutingEditMessagePlanSettingsPage,
   RoutingGetReadyToMovePage,
   RoutingInvalidMessagePlanPage,
   RoutingMessagePlanCampaignIdRequiredPage,
@@ -99,6 +98,7 @@ const protectedPages = [
   RoutingPreviewOtherLanguageLetterTemplatePage,
   RoutingPreviewSmsTemplatePage,
   RoutingPreviewStandardLetterTemplatePage,
+  RoutingRenameMessagePlanPage,
   RoutingReviewAndMoveToProductionPage,
   TemplateMgmtChoosePage,
   TemplateMgmtChoosePrintingAndPostagePage,
@@ -126,8 +126,8 @@ const protectedPages = [
   TemplateMgmtPreviewSubmittedLetterPage,
   TemplateMgmtPreviewSubmittedNhsAppPage,
   TemplateMgmtPreviewSubmittedSmsPage,
-  TemplateMgmtRequestProofPage,
   TemplateMgmtRequestADigitalProofPage,
+  TemplateMgmtRequestProofPage,
   TemplateMgmtReviewAndApproveLetterTemplatePage,
   TemplateMgmtSubmitEmailPage,
   TemplateMgmtSubmitLetterPage,
